requirements.txt
tensorflow-gpu==1.8.0
tensorflow_hub==0.1.1
